src.zip
Implement and apply one or more of the classical search algorithms that we have discussed
in class to solve a range of problems in the domain, and comment on their performance. A*
should be among them. you need to use at least three such as A*, Greedy,
and Uniform-Cost search. Make sure that the implementation is independent of the specific
domain. This means that the same algorithm should be able to solve other variants of the
problem (e.g. the polygons in the plane are different).